clean(#52): revert move of definitions.py
