+++ main.py
@@ -15,7 +15,7 @@ from stylometry.corenlp import connect_corenlp
 from stylometry.logistic_regression import predict_author
 from misc.tem_helpers import get_te_graph, get_tecm
 from train_tem_metrics import predict_from_tecm
-from misc.definitions import ROOT_DIR
+from definitions import ROOT_DIR
 
 def __models_thread():
     proc = subprocess.run([os.path.join(ROOT_DIR, 'prepare_models')], capture_output=True)
