+++ misc/mock_database.py
@@ -1,4 +1,4 @@
-from misc.definitions import DATABASE_AUTHORS_PATH, DATABASE_FILES_PATH, DATABASE_GEN_PATH, DATABASE_TEST_PATH, DATABASE_GERMAN_PATH, DATABASE_GERMAN_TEST_PATH
+from definitions import DATABASE_AUTHORS_PATH, DATABASE_FILES_PATH, DATABASE_GEN_PATH, DATABASE_TEST_PATH, DATABASE_GERMAN_PATH, DATABASE_GERMAN_TEST_PATH
 import yaml, os
