+++ train_tem_metrics.py
@@ -6,7 +6,7 @@ import numpy.typing as npt
 import pandas as pd
 from sklearn.linear_model import LogisticRegression
 from sklearn.model_selection import cross_val_score, RepeatedStratifiedKFold
-from misc.definitions import TEMMETRICS_DIR, TEM_PARAMS
+from definitions import TEMMETRICS_DIR, TEM_PARAMS
 from misc.mock_database import DatabaseAuthorship, DatabaseGenArticles, GermanDatabase
 from misc.tem_helpers import get_tecm
 from misc.nlp_helpers import handle_nltk_download
